Splitting ParametricParticle into three files.
Moving stream operation into MonteCarlo.

Adding energy and energy weight to MeshParticle. Updating the distance
functions.

Sketching out the needed functionality in MonteCarlo.

Changing some names to protect the innocent.

--HG--
rename : src/Particle/Test/ParametricParticle_test.hs => src/Particle/Test/Particle_test.hs

+++ src/NormalizedValues.hs
@@ -37,7 +37,7 @@ class Mag a where
   normalize  :: a -> Normalized a
   magnitude  :: a -> Double
   magnitude2 :: a -> Double  -- ^ Square of the magnitude.
-  magnitude2 x = (magnitude x) ^ 2
+  magnitude2 x = (magnitude x) ^ (2::Integer)
 
 instance Mag Double where
   normalize  d = Normalized $ if d < 0 then -1 else 1  -- Right biased.
